Set env-vars via unicorn config file
This introduces a new way to propagate env-vars to the unicorn processes
via its config file. The reason for this is to make sure that the new
env-vars gets propagated on deploy.

Previous to this the env-vars got propagated only on start, this will
make sure we regenerate the config file and that way we can propagate
the envvars on a regular deploy.